agents/kennedy/requirements.txt
# Kennedy — Media Director Dependencies
python-telegram-bot>=20.0
openai>=1.0.0           # xAI Grok uses OpenAI-compatible API
python-dotenv>=1.0.0
requests>=2.31.0
tweepy>=4.14.0          # X/Twitter API
praw>=7.7.0             # Reddit API
google-api-python-client>=2.100.0  # YouTube Data API
google-analytics-data>=0.18.0      # Google Analytics 4 API
